Add footer to panel
