+++ libs/designer-ui/src/lib/workflowparameters/index.ts
@@ -1,2 +1,3 @@
 export * from './workflowparameter';
 export * from './workflowparameters';
+export * from './workflowparametersFooter';
