+++ libs/designer-ui/src/lib/workflowparameters/workflowparameters.less
@@ -78,11 +78,6 @@
     padding: 30px 50px;
   }
 
-  .msla-workflow-parameters-link {
-    position: absolute;
-    bottom: 20px;
-  }
-
   .msla-workflow-parameters-link-text {
     padding-right: 5px;
   }
